feat(nvim): Add vim-plug for plugin management
This commit adds support for vim-plug, a plugin manager for Neovim. If vim-plug is not already installed, it will be downloaded and installed automatically. Once installed, the plugins listed in the configuration file will be installed and configured.

This change aims to simplify the plugin management process and make it easier to add, remove, and update plugins in Neovim.