Add script for installing and setting up Vim plugins
This commit adds a Lua script that installs vim-plug if it's not already installed and then proceeds to install the necessary plugins specified in the script. After the installation, it reloads the configuration file and sets up additional plugin-specific options. The script is required by init.lua to prepare for using the plugins.