feat: Add plugins folder to runtimepath
The plugins folder is appended to the runtimepath in order to enable Neovim to load plugins from the folder. This change ensures that any plugins placed in the plugins folder will be recognized and loaded by Neovim. It simplifies the process of adding and managing plugins, providing greater flexibility and customization options for the user. No issues are associated with this change.